Blend-groups: split out the attributes from the basic exploration of blend groups.
