+++ examples/03A-blend-group-attributes/002-blend-stroke-width.js
@@ -1,26 +1,23 @@
 /*
-Example: 011-blend-dash-pattern
+Example: 002-blend-stroke-width
 
 Purpose:
 
-To demonstrate how dash patterns are interpolated with blend groups.
+To demonstrate how stroke widths are interpolated with blend groups.
 
 The file deals with five cases
 
-1. Attempt to blend from empty pattern to dashed pattern offset 0
-2. Attempt to blend from empty pattern to empty pattern
-3. Attempt to blend from dashed pattern offset 0 to empty pattern
-4. Attempt to blend from dashed pattern offset 0 to dashed pattern offset 5000
-5. Attempt to blend from dashed pattern offset 0 to dashed pattern offset 0
+1. Attempt to blend from width 6000 to width 3000
+2. Attempt to blend from width 6000 to width none
+3. Attempt to blend from width none to width 3000
+4. Attempt to blend from width none to width none
+5. Attempt to blend from width 6000 to width 6000
 
-The behaviour here is very odd.
+Important point to note:
 
-When a join type is specified the viewport is filled with blue (the stroke colour).
+If the join style is missing from the file then the interpolated shapes are not stroked.
 
-When a join type isn't specified everything renders as expected,
-except of course the intermediate shapes aren't stroked.
-
-Case 2 is the only case that render properly when on its own within the file.
+If the join style is present then the attribute is interpolated as expected.
  */
 
 const {
@@ -31,11 +28,11 @@ const {
 
 const {
   STROKE_COLOUR_BLUE,
+  STROKE_WIDTH_NONE,
+  STROKE_WIDTH_3000,
   STROKE_WIDTH_6000,
   WINDING_RULE_EVEN_ODD,
   DASH_PATTERN_EMPTY,
-  DASH_PATTERN_OFFSET_0,
-  DASH_PATTERN_OFFSET_5000,
   JOIN_BEVEL,
   WORK_AREA,
   FILL_FLAT_BLACK_30,
@@ -47,45 +44,45 @@ const {
 
 const { createClosedSquare } = require('../path-creators');
 
-const createSimpleBlendGroup = require('./shared');
+const createSimpleBlendGroup = require('../simple-blend-group');
 
 const GROUP_0 = createSimpleBlendGroup(
   createClosedSquare(100_000, 100_000, 100_000),
-  DASH_PATTERN_EMPTY,
+  STROKE_WIDTH_6000,
   createClosedSquare(500_000, 100_000, 100_000),
-  DASH_PATTERN_OFFSET_0,
+  STROKE_WIDTH_3000,
   3,
 );
 
 const GROUP_1 = createSimpleBlendGroup(
   createClosedSquare(100_000, 250_000, 100_000),
-  DASH_PATTERN_EMPTY,
+  STROKE_WIDTH_6000,
   createClosedSquare(500_000, 250_000, 100_000),
-  DASH_PATTERN_EMPTY,
+  STROKE_WIDTH_NONE,
   3,
 );
 
 const GROUP_2 = createSimpleBlendGroup(
   createClosedSquare(100_000, 400_000, 100_000),
-  DASH_PATTERN_OFFSET_0,
+  STROKE_WIDTH_NONE,
   createClosedSquare(500_000, 400_000, 100_000),
-  DASH_PATTERN_EMPTY,
+  STROKE_WIDTH_3000,
   3,
 );
 
 const GROUP_3 = createSimpleBlendGroup(
   createClosedSquare(100_000, 550_000, 100_000),
-  DASH_PATTERN_OFFSET_0,
+  STROKE_WIDTH_NONE,
   createClosedSquare(500_000, 550_000, 100_000),
-  DASH_PATTERN_OFFSET_5000,
+  STROKE_WIDTH_NONE,
   3,
 );
 
 const GROUP_4 = createSimpleBlendGroup(
   createClosedSquare(100_000, 700_000, 100_000),
-  DASH_PATTERN_OFFSET_0,
+  STROKE_WIDTH_6000,
   createClosedSquare(500_000, 700_000, 100_000),
-  DASH_PATTERN_OFFSET_0,
+  STROKE_WIDTH_6000,
   3,
 );
