+++ examples/03A-blend-group-attributes/006-blend-join-style.js
@@ -1,26 +1,20 @@
 /*
-Example: 002-blend-stroke-colour
+Example: 006-blend-join-style
 
 Purpose:
 
-To demonstrate how stroke colours are interpolated with blend groups.
+To demonstrate how join styles are interpolated with blend groups.
 
 The file deals with five cases
 
-1. Attempt to blend from blue to red stroke
-2. Attempt to blend from blue to transparent stroke
-3. Attempt to blend from transparent to red stroke
-4. Attempt to blend from transparent to transparent stroke
-5. Attempt to blend from blue to blue stroke.
+1. Attempt to blend from round join to round join
+2. Attempt to blend from round join to bevel join
+3. Attempt to blend from round join to mitre join
+4. Attempt to blend from mitre join to round join
+5. Attempt to blend from bevel join to round join
 
-Important point to note:
-
-If the join style is missing from the file then the interpolated shapes are not stroked.
-
-If the join style is present then the attribute is interpolated as expected.
-
-When interpolating to/from transparent the colours are mixed with white.
-
+The rule appears to be that before the halfway point use the start join and
+beyond that the end join.
  */
 
 const {
@@ -30,13 +24,13 @@ const {
 } = require('../../src').Artworks;
 
 const {
-  STROKE_COLOUR_TRANSPARENT,
   STROKE_COLOUR_BLUE,
-  STROKE_COLOUR_RED,
   STROKE_WIDTH_6000,
   WINDING_RULE_EVEN_ODD,
   DASH_PATTERN_EMPTY,
+  JOIN_ROUND,
   JOIN_BEVEL,
+  JOIN_MITRE,
   WORK_AREA,
   FILL_FLAT_BLACK_30,
 } = require('../shared-objects');
@@ -47,45 +41,45 @@ const {
 
 const { createClosedSquare } = require('../path-creators');
 
-const createSimpleBlendGroup = require('./shared');
+const createSimpleBlendGroup = require('../simple-blend-group');
 
 const GROUP_0 = createSimpleBlendGroup(
   createClosedSquare(100_000, 100_000, 100_000),
-  STROKE_COLOUR_BLUE,
+  JOIN_ROUND,
   createClosedSquare(500_000, 100_000, 100_000),
-  STROKE_COLOUR_RED,
+  JOIN_ROUND,
   3,
 );
 
 const GROUP_1 = createSimpleBlendGroup(
   createClosedSquare(100_000, 250_000, 100_000),
-  STROKE_COLOUR_BLUE,
+  JOIN_ROUND,
   createClosedSquare(500_000, 250_000, 100_000),
-  STROKE_COLOUR_TRANSPARENT,
+  JOIN_BEVEL,
   3,
 );
 
 const GROUP_2 = createSimpleBlendGroup(
   createClosedSquare(100_000, 400_000, 100_000),
-  STROKE_COLOUR_TRANSPARENT,
+  JOIN_ROUND,
   createClosedSquare(500_000, 400_000, 100_000),
-  STROKE_COLOUR_RED,
+  JOIN_MITRE,
   3,
 );
 
 const GROUP_3 = createSimpleBlendGroup(
   createClosedSquare(100_000, 550_000, 100_000),
-  STROKE_COLOUR_TRANSPARENT,
+  JOIN_MITRE,
   createClosedSquare(500_000, 550_000, 100_000),
-  STROKE_COLOUR_TRANSPARENT,
+  JOIN_ROUND,
   3,
 );
 
 const GROUP_4 = createSimpleBlendGroup(
   createClosedSquare(100_000, 700_000, 100_000),
-  STROKE_COLOUR_BLUE,
+  JOIN_BEVEL,
   createClosedSquare(500_000, 700_000, 100_000),
-  STROKE_COLOUR_BLUE,
+  JOIN_ROUND,
   3,
 );
 
@@ -93,8 +87,8 @@ module.exports = createArtworks(
   List.of(WINDING_RULE_EVEN_ODD),
   List.of(DASH_PATTERN_EMPTY),
   List.of(FILL_FLAT_BLACK_30),
+  List.of(STROKE_COLOUR_BLUE),
   List.of(STROKE_WIDTH_6000),
-  List.of(JOIN_BEVEL),
   GROUP_0,
   GROUP_1,
   GROUP_2,
