+++ examples/03A-blend-group-attributes/008-blend-start-line-cap.js
@@ -1,19 +1,19 @@
 /*
-Example: 012-blend-end-marker
+Example: 008-blend-start-line-cap
 
 Purpose:
 
-To demonstrate how end markers are interpolated with blend groups.
+To demonstrate how start line caps are interpolated with blend groups.
 
 The file deals with five cases
 
-1. Attempt to blend from triangle 4,4 to arrowhead 4,6
-2. Attempt to blend from triangle 4,4 to circle 6,4
-3. Attempt to blend from triangle 4,4 to triangle 4,4
-4. Attempt to blend from triangle 4,4 to arrow tail 4,4
-5. Attempt to blend from arrow tail 4,4 to triangle 4,4
+1. Attempt to blend from butt cap to round cap
+2. Attempt to blend from butt cap to square cap
+3. Attempt to blend from butt cap to butt cap
+4. Attempt to blend from butt cap to triangle cap
+5. Attempt to blend from triangle cap to butt cap
 
-The end markers are only applied for the start and end states.
+The start caps are only applied for the start and end states.
 
 The intermediate states draw a closed contour.
  */
@@ -30,11 +30,11 @@ const {
   WINDING_RULE_EVEN_ODD,
   DASH_PATTERN_EMPTY,
   JOIN_BEVEL,
-  START_MARKER_NONE,
-  END_MARKER_TRIANGLE_W4_H4,
-  END_MARKER_ARROW_HEAD_W4_H6,
-  END_MARKER_CIRCLE_W6_H4,
-  END_MARKER_ARROW_TAIL_W4_H4,
+  END_CAP_BUTT,
+  START_CAP_BUTT,
+  START_CAP_ROUND,
+  START_CAP_SQUARE,
+  START_CAP_TRIANGLE_W2_H4,
   WORK_AREA,
   FILL_FLAT_BLACK_30,
 } = require('../shared-objects');
@@ -45,45 +45,45 @@ const {
 
 const { createOpenInvertedV } = require('../path-creators');
 
-const createSimpleBlendGroup = require('./shared');
+const createSimpleBlendGroup = require('../simple-blend-group');
 
 const GROUP_0 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 100_000, 100_000),
-  END_MARKER_TRIANGLE_W4_H4,
+  START_CAP_BUTT,
   createOpenInvertedV(500_000, 100_000, 100_000),
-  END_MARKER_ARROW_HEAD_W4_H6,
+  START_CAP_ROUND,
   3,
 );
 
 const GROUP_1 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 250_000, 100_000),
-  END_MARKER_TRIANGLE_W4_H4,
+  START_CAP_BUTT,
   createOpenInvertedV(500_000, 250_000, 100_000),
-  END_MARKER_CIRCLE_W6_H4,
+  START_CAP_SQUARE,
   3,
 );
 
 const GROUP_2 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 400_000, 100_000),
-  END_MARKER_TRIANGLE_W4_H4,
+  START_CAP_BUTT,
   createOpenInvertedV(500_000, 400_000, 100_000),
-  END_MARKER_TRIANGLE_W4_H4,
+  START_CAP_BUTT,
   3,
 );
 
 const GROUP_3 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 550_000, 100_000),
-  END_MARKER_TRIANGLE_W4_H4,
+  START_CAP_BUTT,
   createOpenInvertedV(500_000, 550_000, 100_000),
-  END_MARKER_ARROW_TAIL_W4_H4,
+  START_CAP_TRIANGLE_W2_H4,
   3,
 );
 
 const GROUP_4 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 700_000, 100_000),
-  END_MARKER_ARROW_TAIL_W4_H4,
+  START_CAP_TRIANGLE_W2_H4,
   createOpenInvertedV(500_000, 700_000, 100_000),
-  END_MARKER_TRIANGLE_W4_H4,
+  START_CAP_BUTT,
   3,
 );
 
@@ -93,7 +93,7 @@ module.exports = createArtworks(
   List.of(FILL_FLAT_BLACK_30),
   List.of(STROKE_COLOUR_BLUE),
   List.of(STROKE_WIDTH_6000),
-  List.of(START_MARKER_NONE),
+  List.of(END_CAP_BUTT),
   List.of(JOIN_BEVEL),
   GROUP_0,
   GROUP_1,
