+++ examples/03A-blend-group-attributes/010-blend-dash-pattern.js
@@ -1,20 +1,26 @@
 /*
-Example: 007-blend-join-style
+Example: 010-blend-dash-pattern
 
 Purpose:
 
-To demonstrate how join styles are interpolated with blend groups.
+To demonstrate how dash patterns are interpolated with blend groups.
 
 The file deals with five cases
 
-1. Attempt to blend from round join to round join
-2. Attempt to blend from round join to bevel join
-3. Attempt to blend from round join to mitre join
-4. Attempt to blend from mitre join to round join
-5. Attempt to blend from bevel join to round join
+1. Attempt to blend from empty pattern to dashed pattern offset 0
+2. Attempt to blend from empty pattern to empty pattern
+3. Attempt to blend from dashed pattern offset 0 to empty pattern
+4. Attempt to blend from dashed pattern offset 0 to dashed pattern offset 5000
+5. Attempt to blend from dashed pattern offset 0 to dashed pattern offset 0
 
-The rule appears to be that before the halfway point use the start join and
-beyond that the end join.
+The behaviour here is very odd.
+
+When a join type is specified the viewport is filled with blue (the stroke colour).
+
+When a join type isn't specified everything renders as expected,
+except of course the intermediate shapes aren't stroked.
+
+Case 2 is the only case that render properly when on its own within the file.
  */
 
 const {
@@ -28,9 +34,9 @@ const {
   STROKE_WIDTH_6000,
   WINDING_RULE_EVEN_ODD,
   DASH_PATTERN_EMPTY,
-  JOIN_ROUND,
+  DASH_PATTERN_OFFSET_0,
+  DASH_PATTERN_OFFSET_5000,
   JOIN_BEVEL,
-  JOIN_MITRE,
   WORK_AREA,
   FILL_FLAT_BLACK_30,
 } = require('../shared-objects');
@@ -41,45 +47,45 @@ const {
 
 const { createClosedSquare } = require('../path-creators');
 
-const createSimpleBlendGroup = require('./shared');
+const createSimpleBlendGroup = require('../simple-blend-group');
 
 const GROUP_0 = createSimpleBlendGroup(
   createClosedSquare(100_000, 100_000, 100_000),
-  JOIN_ROUND,
+  DASH_PATTERN_EMPTY,
   createClosedSquare(500_000, 100_000, 100_000),
-  JOIN_ROUND,
+  DASH_PATTERN_OFFSET_0,
   3,
 );
 
 const GROUP_1 = createSimpleBlendGroup(
   createClosedSquare(100_000, 250_000, 100_000),
-  JOIN_ROUND,
+  DASH_PATTERN_EMPTY,
   createClosedSquare(500_000, 250_000, 100_000),
-  JOIN_BEVEL,
+  DASH_PATTERN_EMPTY,
   3,
 );
 
 const GROUP_2 = createSimpleBlendGroup(
   createClosedSquare(100_000, 400_000, 100_000),
-  JOIN_ROUND,
+  DASH_PATTERN_OFFSET_0,
   createClosedSquare(500_000, 400_000, 100_000),
-  JOIN_MITRE,
+  DASH_PATTERN_EMPTY,
   3,
 );
 
 const GROUP_3 = createSimpleBlendGroup(
   createClosedSquare(100_000, 550_000, 100_000),
-  JOIN_MITRE,
+  DASH_PATTERN_OFFSET_0,
   createClosedSquare(500_000, 550_000, 100_000),
-  JOIN_ROUND,
+  DASH_PATTERN_OFFSET_5000,
   3,
 );
 
 const GROUP_4 = createSimpleBlendGroup(
   createClosedSquare(100_000, 700_000, 100_000),
-  JOIN_BEVEL,
+  DASH_PATTERN_OFFSET_0,
   createClosedSquare(500_000, 700_000, 100_000),
-  JOIN_ROUND,
+  DASH_PATTERN_OFFSET_0,
   3,
 );
 
@@ -89,6 +95,7 @@ module.exports = createArtworks(
   List.of(FILL_FLAT_BLACK_30),
   List.of(STROKE_COLOUR_BLUE),
   List.of(STROKE_WIDTH_6000),
+  List.of(JOIN_BEVEL),
   GROUP_0,
   GROUP_1,
   GROUP_2,
