+++ examples/03A-blend-group-attributes/011-blend-end-marker.js
@@ -1,9 +1,9 @@
 /*
-Example: 013-blend-start-marker
+Example: 011-blend-end-marker
 
 Purpose:
 
-To demonstrate how start markers are interpolated with blend groups.
+To demonstrate how end markers are interpolated with blend groups.
 
 The file deals with five cases
 
@@ -13,7 +13,7 @@ The file deals with five cases
 4. Attempt to blend from triangle 4,4 to arrow tail 4,4
 5. Attempt to blend from arrow tail 4,4 to triangle 4,4
 
-The start markers are only applied for the start and end states.
+The end markers are only applied for the start and end states.
 
 The intermediate states draw a closed contour.
  */
@@ -30,11 +30,11 @@ const {
   WINDING_RULE_EVEN_ODD,
   DASH_PATTERN_EMPTY,
   JOIN_BEVEL,
-  END_MARKER_NONE,
-  START_MARKER_TRIANGLE_W4_H4,
-  START_MARKER_ARROW_HEAD_W4_H6,
-  START_MARKER_CIRCLE_W6_H4,
-  START_MARKER_ARROW_TAIL_W4_H4,
+  START_MARKER_NONE,
+  END_MARKER_TRIANGLE_W4_H4,
+  END_MARKER_ARROW_HEAD_W4_H6,
+  END_MARKER_CIRCLE_W6_H4,
+  END_MARKER_ARROW_TAIL_W4_H4,
   WORK_AREA,
   FILL_FLAT_BLACK_30,
 } = require('../shared-objects');
@@ -45,45 +45,45 @@ const {
 
 const { createOpenInvertedV } = require('../path-creators');
 
-const createSimpleBlendGroup = require('./shared');
+const createSimpleBlendGroup = require('../simple-blend-group');
 
 const GROUP_0 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 100_000, 100_000),
-  START_MARKER_TRIANGLE_W4_H4,
+  END_MARKER_TRIANGLE_W4_H4,
   createOpenInvertedV(500_000, 100_000, 100_000),
-  START_MARKER_ARROW_HEAD_W4_H6,
+  END_MARKER_ARROW_HEAD_W4_H6,
   3,
 );
 
 const GROUP_1 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 250_000, 100_000),
-  START_MARKER_TRIANGLE_W4_H4,
+  END_MARKER_TRIANGLE_W4_H4,
   createOpenInvertedV(500_000, 250_000, 100_000),
-  START_MARKER_CIRCLE_W6_H4,
+  END_MARKER_CIRCLE_W6_H4,
   3,
 );
 
 const GROUP_2 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 400_000, 100_000),
-  START_MARKER_TRIANGLE_W4_H4,
+  END_MARKER_TRIANGLE_W4_H4,
   createOpenInvertedV(500_000, 400_000, 100_000),
-  START_MARKER_TRIANGLE_W4_H4,
+  END_MARKER_TRIANGLE_W4_H4,
   3,
 );
 
 const GROUP_3 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 550_000, 100_000),
-  START_MARKER_TRIANGLE_W4_H4,
+  END_MARKER_TRIANGLE_W4_H4,
   createOpenInvertedV(500_000, 550_000, 100_000),
-  START_MARKER_ARROW_TAIL_W4_H4,
+  END_MARKER_ARROW_TAIL_W4_H4,
   3,
 );
 
 const GROUP_4 = createSimpleBlendGroup(
   createOpenInvertedV(100_000, 700_000, 100_000),
-  START_MARKER_ARROW_TAIL_W4_H4,
+  END_MARKER_ARROW_TAIL_W4_H4,
   createOpenInvertedV(500_000, 700_000, 100_000),
-  START_MARKER_TRIANGLE_W4_H4,
+  END_MARKER_TRIANGLE_W4_H4,
   3,
 );
 
@@ -93,7 +93,7 @@ module.exports = createArtworks(
   List.of(FILL_FLAT_BLACK_30),
   List.of(STROKE_COLOUR_BLUE),
   List.of(STROKE_WIDTH_6000),
-  List.of(END_MARKER_NONE),
+  List.of(START_MARKER_NONE),
   List.of(JOIN_BEVEL),
   GROUP_0,
   GROUP_1,
