+++ examples/03A-blend-group/001-blend-step-variants.js
@@ -33,7 +33,7 @@ const {
 
 const { createClosedSquare } = require('../path-creators');
 
-const createSimpleBlendGroup = require('./shared');
+const createSimpleBlendGroup = require('../simple-blend-group');
 
 function createSimpleBlendGroups() {
   const groups = [];
