+++ scripts/examples.js
@@ -22,7 +22,8 @@ const EXAMPLE_DIRECTORIES = [
   '02C-rectangle',
   '034-ellipse',
   '035-rounded-rectangle',
-  '03A-blend-groups',
+  '03A-blend-group',
+  '03A-blend-group-attributes',
   '03E-start-markers',
   '03F-end-markers',
   '100-smallest-file',
